move dependency solver to new directory
