+++ nuget/helpers/lib/NuGetUpdater/NuGetUpdater.Core/DependencySolver/IDependencySolver.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace NuGetUpdater.Core.Updater.FileWriters;
+namespace NuGetUpdater.Core.DependencySolver;
 
 public interface IDependencySolver
 {
